move mapStyles under map
