+++ src/map/style.js
